renamed `FirestoreFunction` to `CloudFunction`, added `acceptLinkRequest`, using `UserType` enum in `joinRequest`
